requirements.txt
python-telegram-bot>=20.0
requests==2.28.1
beautifulsoup4==4.12.2
